Structural Changes to Platform Documentation; One Routing Tier, Two Subdomains
Major Decision: We are going to describe our Routing Tier as being just one Routing Tier with two options via the two subdomains. This is for clarity, and to emphasize that both options have distinct possibilities and benefits, and can also be used in combination. Many of the following changes are a result of this logic, and should balance out the perceived value of the two subdomains.

Structural Changes —
- “Provided Subdomains” and  “Routing Tier” need to be closer together, logically. Thus the Scaling section has been shifted below Routing Tier.
- Websockets has been given its own section, to put it at an importance level equal to Caching.
Domains subsection under “Routing Tier” has been removed because it repeats information in “Provided Domains”; Elastic Addresses subsection shifted to “Provided Domains”.
- Failover and Health Checker subsections are included under new H3 header “High Availability”.
- Timeouts tables moved up to “Reverse Proxy Timeouts” Section

Language Changes —
- All mentions of having two different routing tiers have been changed to one routing tier with two options.
- The documentation is meant to be a “static” document, which is clear to read now, a month, or a year from now. This means all mentions of “beta”, “announcing”, or “new” have been removed.

Conflicts:
	Platform Documentation.md