Add worker.restart to the dotCloud documentation
Update the application code by pushing and redeploying to the platform doesn't
restart the running workers automatically. The information how to
restart those workers manually is missing in the worker documentation.
In addition there is a note missing for the deploy command option
`--restart-workers`.

Add the command and notes to the documentation to provide the necessary
information.