Add GCE specific information to the Sendgrid documentation
dotCloud operates on GCE and provides Sendgrid as a mail delivery
solution on the platform. The Google Compute Engine blocks the standard
outgoing ports for SMTP (i.e., 25, 465, 587) and uses the SMTP-Port:
2525 instead. Therefore, if the SMTP-Port is not set to 2525 in the
application, users are not able to send mails via dotCloud.

Add the missing information to improve the documentation