Update the Flask documentation
Due to the outdated versions of the used Flask version.
The documentation needed a rework to correct the dependency.
Add the missing Application code and explanation.

Another need is to rework on the documentation Layouts and texts
to uniform the style of the guides. More Pull Requests will follow-up
to solve the guide update issue.